수리접수 오류수정

+++ src/main/java/com/surit/fixer/estimate/service/EstimateService.java
@@ -9,11 +9,11 @@ import com.surit.fixer.estimate.model.dto.EstimateForm;
 public interface EstimateService {
 
 	/** 견적 작성 화면에 띄울 접수 정보 (볼 수 없는 접수면 예외) */
-	RequestDTO getTargetRequest(int fixerNo, long requestId);
+	RequestDTO getTargetRequest(Long fixerNo, Long requestId);
 
 	/** 견적 제출 */
-	void submit(int fixerNo, EstimateForm form);
+	void submit(Long fixerNo, EstimateForm form);
 
 	/** 내가 낸 견적 목록 */
-	List<EstimateDTO> getMyEstimates(int fixerNo);
+	List<EstimateDTO> getMyEstimates(Long fixerNo);
 }
