+++ src/main/java/com/surit/fixer/job/model/mapper/JobMapper.java
@@ -15,12 +15,12 @@ public interface JobMapper {
 	 *
 	 * @param statusCode 상태 필터 (없으면 null → 전체)
 	 */
-	List<JobDTO> selectMyJobs(@Param("fixerNo") int fixerNo,
+	List<JobDTO> selectMyJobs(@Param("fixerNo") Long fixerNo,
 	                          @Param("statusCode") String statusCode);
 
 	/** 내 작업 상세 (내 작업이 아니면 null) */
-	JobDTO selectMyJob(@Param("fixerNo") int fixerNo,
-	                   @Param("requestId") long requestId);
+	JobDTO selectMyJob(@Param("fixerNo") Long fixerNo,
+	                   @Param("requestId") Long requestId);
 
 	/**
 	 * 수리 완료 처리 (REQ_03 → REQ_04).
@@ -31,6 +31,6 @@ public interface JobMapper {
 	 *
 	 * @return 1 이면 성공, 0 이면 조건 불일치
 	 */
-	int completeJob(@Param("fixerNo") int fixerNo,
-	                @Param("requestId") long requestId);
+	Long completeJob(@Param("fixerNo") Long fixerNo,
+	                @Param("requestId") Long requestId);
 }
