+++ src/main/java/com/surit/fixer/job/service/JobService.java
@@ -7,11 +7,11 @@ import com.surit.fixer.job.model.dto.JobDTO;
 public interface JobService {
 
 	/** 내 작업 목록 (statusCode 가 null 이면 전체) */
-	List<JobDTO> getMyJobs(int fixerNo, String statusCode);
+	List<JobDTO> getMyJobs(Long fixerNo, String statusCode);
 
 	/** 내 작업 상세 (내 작업이 아니면 예외) */
-	JobDTO getMyJob(int fixerNo, long requestId);
+	JobDTO getMyJob(Long fixerNo, Long requestId);
 
 	/** 수리 완료 처리 */
-	void complete(int fixerNo, long requestId);
+	void complete(Long fixerNo, Long requestId);
 }
