+++ src/main/java/com/surit/fixer/verify/model/mapper/FixerMapper.java
@@ -14,10 +14,10 @@ public interface FixerMapper {
 	// ---------- 조회 ----------
 
 	/** 기사 프로필 (신청 이력이 있는지 확인용). 없으면 null */
-	FixerProfileDTO selectFixerProfile(int userNo);
+	FixerProfileDTO selectFixerProfile(Long userNo);
 
 	/** 기사가 등록한 자격증 목록 (재신청 시 기존 파일 삭제용) */
-	List<FixerLicenseDTO> selectLicensesByUserNo(int userNo);
+	List<FixerLicenseDTO> selectLicensesByUserNo(Long userNo);
 
 	// ---------- 등록 / 수정 ----------
 
@@ -26,15 +26,15 @@ public interface FixerMapper {
 
 	int insertFixerLicense(FixerLicenseDTO license);
 
-	int insertFixerRegion(@Param("userNo") int userNo,
+	int insertFixerRegion(@Param("userNo") Long userNo,
 	                      @Param("regionCode") String regionCode);
 
-	int insertFixerCategory(@Param("userNo") int userNo,
+	int insertFixerCategory(@Param("userNo") Long userNo,
 	                        @Param("categoryCode") String categoryCode);
 
 	// ---------- 삭제 (재신청 시 기존 데이터 정리) ----------
 
-	int deleteLicensesByUserNo(int userNo);
-	int deleteRegionsByUserNo(int userNo);
-	int deleteCategoriesByUserNo(int userNo);
+	int deleteLicensesByUserNo(Long userNo);
+	int deleteRegionsByUserNo(Long userNo);
+	int deleteCategoriesByUserNo(Long userNo);
 }
